Unify JS into miso.js, refactor tests
- Combines all 4 js files into one.
- Updates tests to reflect this
- diff.tests.js -> miso.spec.js
