+++ tests/miso.spec.js
@@ -1,6 +1,8 @@
-const diff = require('./diff');
-const isomorphic = require('./isomorphic.js');
-const jsdom = require('jsdom');
+/* custom js imports */
+const miso     = require('./miso');
+
+/* third party imports */
+const jsdom    = require('jsdom');
 
 function vcomp(mount, unmount, props, css, children, ref, oc, od, bd, key) {
   return {
@@ -85,12 +87,12 @@ function vtextKeyed(txt, key) {
 
 
 // base case
-test('Should be null when diffing two null virtual DOMs', () => {
+test('Should be null when miso.diffing two null virtual DOMs', () => {
   const document = new jsdom.JSDOM().window.document;
   const body = document.body;
   var c = null;
   n = null;
-  window['diff'](c, n, body, document);
+  miso.diff(c, n, body, document);
   expect(body.childNodes.length).toBe(0);
 });
 
@@ -101,41 +103,41 @@ test('Should create a new text node', () => {
     type: 'vtext',
     text: 'foo'
   };
-  window['diff'](null, newNode, body, doc);
+  miso.diff(null, newNode, body, doc);
   expect(newNode.domRef.wholeText).toBe('foo');
 });
 
-test('Should window diff two identical text nodes', () => {
+test('Should window miso.diff two identical text nodes', () => {
   const doc = new jsdom.JSDOM().window.document;
   const body = doc.body;
   var currentNode = {
     type: 'vtext',
     text: 'foo'
   };
-  window['diff'](null, currentNode, body, doc);
+  miso.diff(null, currentNode, body, doc);
   expect(currentNode.domRef.wholeText).toBe('foo');
   var newNode = {
     type: 'vtext',
     text: 'foo'
   };
-  window['diff'](currentNode, newNode, body, doc);
+  miso.diff(currentNode, newNode, body, doc);
   expect('foo').toBe(newNode.domRef.wholeText);
 });
 
-test('Should window diff two window different text nodes', () => {
+test('Should window miso.diff two window miso.different text nodes', () => {
   const doc = new jsdom.JSDOM().window.document;
   const body = doc.body;
   var currentNode = {
     type: 'vtext',
     text: 'foo'
   };
-  window['diff'](null, currentNode, body, doc);
+  miso.diff(null, currentNode, body, doc);
   expect(currentNode.domRef.wholeText).toBe('foo');
   var newNode = {
     type: 'vtext',
     text: 'bar'
   };
-  window['diff'](currentNode, newNode, body, doc);
+  miso.diff(currentNode, newNode, body, doc);
   expect(newNode.domRef.wholeText).toBe('bar')
 });
 
@@ -144,7 +146,7 @@ test('Should create a new DOM node', () => {
   var body = document.body;
   var currentNode = null;
   var newNode = vnode('div', []);
-  window['diff'](currentNode, newNode, body, document);
+  miso.diff(currentNode, newNode, body, document);
   expect(body.children[0]).toBe(newNode.domRef);
 });
 
@@ -153,10 +155,10 @@ test('Should detect duplicate component mounting', () => {
   var body = document.body;
   var mountCount = 0;
   var newComp1 = vcomp((x) => mountCount++, null, { 'data-component-id' : "vcomp-foo"}, {"background-color":"red"}, []);
-  window['diff'](null, newComp1, body, document);
+  miso.diff(null, newComp1, body, document);
   var newComp2 = vcomp((x) => mountCount++, null, { 'data-component-id' : "vcomp-foo"}, {"background-color":"red"}, []);
   var newNode = vnode('div', [newComp2], {}, {}, 'svg');
-  window['diff'](null, newNode, body, document);
+  miso.diff(null, newNode, body, document);
   expect(mountCount).toBe(1);
 });
 
@@ -168,18 +170,18 @@ test('Should mount and unmount a component', () => {
   var mountFunc = function (cb) {
       mountCount++;
       var node = vnode("div", []);
-      diff (null, node, body, document);
+      miso.diff (null, node, body, document);
       cb (node);
   }
   var newNode =
     vcomp(mountFunc, (x) => unmountCount++, { 'id' : "vcomp-foo"}, {"background-color":"red"}, []);
-  window['diff'](null, newNode, body, document);
+  miso.diff(null, newNode, body, document);
   expect(mountCount).toBe(1);
   expect(newNode.children.length).toBe(1);
   expect(newNode.domRef.children.length).toBe(1);
   expect(newNode.domRef.id).toBe("vcomp-foo");
   expect(newNode.domRef.style['background-color']).toBe("red");
-  window['diff'](newNode, null, body, document);
+  miso.diff(newNode, null, body, document);
   expect(unmountCount).toBe(1);
 });
 
@@ -188,7 +190,7 @@ test('Should create an SVG DOM node', () => {
   var body = document.body;
   var currentNode = null;
   var newNode = vnode('div', [], {}, {}, 'svg');
-  window['diff'](currentNode, newNode, body, document);
+  miso.diff(currentNode, newNode, body, document);
   expect(body.children[0]).toBe(newNode.domRef);
 });
 
@@ -197,7 +199,7 @@ test('Should create a MathML DOM node', () => {
   var body = document.body;
   var currentNode = null;
   var newNode = vnode('math', [], {}, {}, 'mathml');
-  window['diff'](currentNode, newNode, body, document);
+  miso.diff(currentNode, newNode, body, document);
   expect(body.children[0]).toBe(newNode.domRef);
 });
 
@@ -208,7 +210,7 @@ test('Should create an SVG DOM node, with href attribute', () => {
   var newNode = vnode('ellipse', [], {
     'href': 'https://google.com'
   }, {}, 'svg');
-  window['diff'](currentNode, newNode, body, document);
+  miso.diff(currentNode, newNode, body, document);
   expect(body.children[0].getAttributeNS("http://www.w3.org/1999/xlink", 'href')).toBe('https://google.com');
 });
 
@@ -219,12 +221,12 @@ test('Should create an SVG DOM node, with href attribute, and change it', () => 
   var newNode = vnode('ellipse', [], {
     'href': 'https://google.com'
   }, {}, 'svg');
-  window['diff'](currentNode, newNode, body, document);
+  miso.diff(currentNode, newNode, body, document);
   expect(body.children[0].getAttributeNS("http://www.w3.org/1999/xlink", 'href')).toBe('https://google.com');
   var newerNode = vnode('ellipse', [], {
     'href': 'https://yahoo.com'
   }, {}, 'svg');
-  window['diff'](newNode, newerNode, body, document);
+  miso.diff(newNode, newerNode, body, document);
   expect(body.children[0].getAttributeNS("http://www.w3.org/1999/xlink", 'href')).toBe('https://yahoo.com');
 });
 
@@ -235,7 +237,7 @@ test('Should create an SVG DOM node, with regular attribute', () => {
   var newNode = vnode('ellipse', [], {
     'rx': '100'
   }, {}, 'svg');
-  window['diff'](currentNode, newNode, body, document);
+  miso.diff(currentNode, newNode, body, document);
   expect(body.children[0].getAttribute('rx')).toBe('100');
 });
 
@@ -246,29 +248,29 @@ test('Should create an SVG DOM node, with regular attribute, and change it', () 
   var newNode = vnode('ellipse', [], {
     'rx': '100'
   }, {}, 'svg');
-  window['diff'](currentNode, newNode, body, document);
+  miso.diff(currentNode, newNode, body, document);
   expect(body.children[0].getAttribute('rx')).toBe('100');
   var newerNode = vnode('ellipse', [], {
     'rx': '200'
   }, {}, 'svg');
-  window['diff'](newNode, newerNode, body, document);
+  miso.diff(newNode, newerNode, body, document);
   expect(body.children[0].getAttribute('rx')).toBe('200');
 });
 
-test('Should replace a Node with a new Node of a window different tag', () => {
+test('Should replace a Node with a new Node of a window miso.different tag', () => {
   var document = new jsdom.JSDOM().window.document;
   var body = document.body;
 
   // populate DOM
   var node = vnode('div', []);
-  window['diff'](null, node, body, document);
+  miso.diff(null, node, body, document);
 
   // Test node was populated
   expect(body.children.length).toBe(1);
 
   // Replace node
   newNode = vnode('a', []);
-  window['diff'](node, newNode, body, document);
+  miso.diff(node, newNode, body, document);
 
   // Test node is removed from DOM
   expect(body.children[0].tagName).toBe('A');
@@ -280,7 +282,7 @@ test('Should create children', () => {
 
   // populate DOM
   var node = vnode('div', [vnode('div', [])]);
-  window['diff'](null, node, body, document);
+  miso.diff(null, node, body, document);
   expect(node.domRef.children.length).toBe(1);
 });
 
@@ -290,16 +292,16 @@ test('Should remove a child', () => {
 
   // populate DOM
   var node = vnode('div', [vnode('div', [])]);
-  window['diff'](null, node, body, document);
+  miso.diff(null, node, body, document);
   expect(node.domRef.children.length).toBe(1);
 
   // populate DOM
   var newNode = vnode('div', []);
-  window['diff'](node, newNode, body, document);
+  miso.diff(node, newNode, body, document);
   expect(node.domRef.children.length).toBe(0);
 });
 
-test('Should Diff attrs of two Components', () => {
+test('Should Miso.Diff attrs of two Components', () => {
   var document = new jsdom.JSDOM().window.document;
   var body = document.body;
 
@@ -312,7 +314,7 @@ test('Should Diff attrs of two Components', () => {
              , { 'background-color' :"red"}
            );
 
-  window['diff'](null, compNode1, body, document);
+  miso.diff(null, compNode1, body, document);
   expect(mountCount).toBe(1);
 
   // Test node was populated
@@ -328,7 +330,7 @@ test('Should Diff attrs of two Components', () => {
              , { 'background-color' : "green"}
            );
 
-  window['diff'](compNode1, compNode2, body, document);
+  miso.diff(compNode1, compNode2, body, document);
   expect(body.childNodes[0].style['background-color']).toBe('green');
 });
 
@@ -338,7 +340,7 @@ test('Should replace Node with Component', () => {
 
   // populate DOM
   var node = vnode('div', []);
-  window['diff'](null, node, body, document);
+  miso.diff(null, node, body, document);
 
   // Test node was populated
   expect(body.childNodes.length).toBe(1);
@@ -346,7 +348,7 @@ test('Should replace Node with Component', () => {
   // Replace node
   var mountCount = 0;
   compNode = vcomp((x) => mountCount++);
-  window['diff'](node, compNode, body, document);
+  miso.diff(node, compNode, body, document);
 
   // Node is removed from DOM, Component is on the DOM
   expect(body.childNodes[0].getAttribute('data-component-id')).toBe('vcomp-id');
@@ -359,7 +361,7 @@ test('Should replace Text with Component', () => {
 
   // populate DOM
   var node = vtext('foo');
-  window['diff'](null, node, body, document);
+  miso.diff(null, node, body, document);
 
   // Test node was populated
   expect(node.domRef.wholeText).toBe('foo')
@@ -368,7 +370,7 @@ test('Should replace Text with Component', () => {
   // Replace node
   var mountCount = 0;
   compNode = vcomp((x) => mountCount++);
-  window['diff'](node, compNode, body, document);
+  miso.diff(node, compNode, body, document);
 
   // Node is removed from DOM, Component is on the DOM
   expect(body.childNodes[0].getAttribute('data-component-id')).toBe('vcomp-id');
@@ -381,14 +383,14 @@ test('Should replace Node with TextNode', () => {
 
   // populate DOM
   var node = vnode('div', []);
-  window['diff'](null, node, body, document);
+  miso.diff(null, node, body, document);
 
   // Test node was populated
   expect(body.childNodes.length).toBe(1);
 
   // Replace node
   textNode = vtext('fooo');
-  window['diff'](node, textNode, body, document);
+  miso.diff(node, textNode, body, document);
 
   // Test node is removed from DOM
   expect(body.childNodes[0].wholeText).toBe('fooo');
@@ -401,7 +403,7 @@ test('Should replace Component with TextNode', () => {
   // populate DOM
   var mountCount = 0, unmountCount = 0;
   var component = vcomp((x) => mountCount++, (x) => unmountCount++);
-  window['diff'](null, component, body, document);
+  miso.diff(null, component, body, document);
 
   // Test component was populated
   expect(body.childNodes.length).toBe(1);
@@ -410,7 +412,7 @@ test('Should replace Component with TextNode', () => {
 
   // Replace component
   textNode = vtext('fooo');
-  window['diff'](component, textNode, body, document);
+  miso.diff(component, textNode, body, document);
 
   // Test node is removed from DOM
   expect(body.childNodes[0].wholeText).toBe('fooo');
@@ -424,7 +426,7 @@ test('Should replace Component with Node', () => {
   // populate DOM
   var mountCount = 0, unmountCount = 0;
   var component = vcomp((x) => mountCount++, (x) => unmountCount++);
-  window['diff'](null, component, body, document);
+  miso.diff(null, component, body, document);
 
   // Test component was populated
   expect(body.childNodes.length).toBe(1);
@@ -433,7 +435,7 @@ test('Should replace Component with Node', () => {
 
   // Replace component
   node = vnode('div', []);
-  window['diff'](component, node, body, document);
+  miso.diff(component, node, body, document);
 
   // Test node is removed from DOM
   expect(body.children[0].tagName).toBe('DIV');
@@ -446,14 +448,14 @@ test('Should replace TextNode with Node', () => {
 
   // populate DOM
   var textNode = vtext('fooo');
-  window['diff'](null, textNode, body, document);
+  miso.diff(null, textNode, body, document);
 
   // Test node was populated
   expect(body.childNodes.length).toBe(1);
 
   // Replace node
   node = vnode('div', []);
-  window['diff'](textNode, node, body, document);
+  miso.diff(textNode, node, body, document);
 
   // Test node is removed from DOM
   expect(body.children[0].tagName).toBe('DIV');
@@ -466,13 +468,13 @@ test('Should remove a DOM node', () => {
   // populate DOM
   var currentNode = null;
   var newNode = vnode('div', []);
-  window['diff'](currentNode, newNode, body, document);
+  miso.diff(currentNode, newNode, body, document);
 
   // Test node was populated
   expect(body.children.length).toBe(1);
 
   // Remove node
-  window['diff'](newNode, null, body, document);
+  miso.diff(newNode, null, body, document);
 
   // Test node is removed from DOM
   expect(body.children.length).toBe(0);
@@ -486,11 +488,11 @@ test('Should create a new property on a DOM node', () => {
   var currentNode = vnode('div', [], {
     'id': 'a'
   });
-  window['diff'](null, currentNode, body, document)
+  miso.diff(null, currentNode, body, document)
   expect(currentNode.domRef['id']).toBe('a');
 });
 
-test('Should skip if window diffing identical properties', () => {
+test('Should skip if window miso.diffing identical properties', () => {
   var document = new jsdom.JSDOM().window.document;
   var body = document.body;
 
@@ -498,12 +500,12 @@ test('Should skip if window diffing identical properties', () => {
   var currentNode = vnode('div', [], {
     'id': 'a'
   });
-  window['diff'](null, currentNode, body, document)
+  miso.diff(null, currentNode, body, document)
 
   var newNode = vnode('div', [], {
     'id': 'a'
   });
-  window['diff'](currentNode, newNode, body, document)
+  miso.diff(currentNode, newNode, body, document)
   expect(currentNode.domRef).toBe(newNode.domRef);
 });
 
@@ -515,7 +517,7 @@ test('Should create a custom attribute on a DOM node', () => {
   var currentNode = vnode('div', [], {
     'lol': 'lol'
   }, {});
-  window['diff'](null, currentNode, body, document)
+  miso.diff(null, currentNode, body, document)
   expect(currentNode.domRef.getAttribute('lol')).toBe('lol');
 });
 
@@ -527,13 +529,13 @@ test('Should change a custom attribute on a DOM node', () => {
   var currentNode = vnode('div', [], {
     'lol': 'lol'
   }, {});
-  window['diff'](null, currentNode, body, document)
+  miso.diff(null, currentNode, body, document)
   expect(currentNode.domRef.getAttribute('lol')).toBe('lol');
 
   var newNode = vnode('div', [], {
     'lol': 'lolz'
   }, {});
-  window['diff'](currentNode, newNode, body, document)
+  miso.diff(currentNode, newNode, body, document)
   expect(currentNode.domRef.getAttribute('lol')).toBe('lolz');
 });
 
@@ -545,12 +547,12 @@ test('Should remove a custom attribute from a DOM node', () => {
   var currentNode = vnode('div', [], {
     'lol': 'lol'
   });
-  window['diff'](null, currentNode, body, document)
+  miso.diff(null, currentNode, body, document)
   expect(currentNode.domRef.getAttribute('lol')).toBe('lol');
 
   // test property change
   var newNode = vnode('div', [], {});
-  window['diff'](currentNode, newNode, body, document);
+  miso.diff(currentNode, newNode, body, document);
   expect(newNode.domRef.getAttribute('lol')).toBe(null);
 });
 
@@ -562,11 +564,11 @@ test('Should remove a property from DOM node', () => {
   var currentNode = vnode('div', [], {
     'id': 'someid'
   });
-  window['diff'](null, currentNode, body, document)
+  miso.diff(null, currentNode, body, document)
 
   // test property change
   var newNode = vnode('div', [], {});
-  window['diff'](currentNode, newNode, body, document);
+  miso.diff(currentNode, newNode, body, document);
   expect(newNode.domRef['id']).toBe('');
 });
 
@@ -578,13 +580,13 @@ test('Should change a property from DOM node', () => {
   var currentNode = vnode('div', [], {
     'id': 'someid'
   });
-  window['diff'](null, currentNode, body, document)
+  miso.diff(null, currentNode, body, document)
 
   // test property change
   var newNode = vnode('div', [], {
     'id': 'foo'
   });
-  window['diff'](currentNode, newNode, body, document);
+  miso.diff(currentNode, newNode, body, document);
   expect(newNode.domRef['id']).toBe('foo');
 });
 
@@ -596,7 +598,7 @@ test('Should create css on a DOM node', () => {
   var newNode = vnode('div', [], {}, {
     'color': 'red'
   });
-  window['diff'](null, newNode, body, document)
+  miso.diff(null, newNode, body, document)
   expect(newNode.domRef.style['color']).toBe('red');
 });
 
@@ -608,11 +610,11 @@ test('Should remove css from DOM node', () => {
   var currentNode = vnode('div', [], {}, {
     'color': 'red'
   });
-  window['diff'](null, currentNode, body, document)
+  miso.diff(null, currentNode, body, document)
 
   // test css change
   var newNode = vnode('div', [], {}, {});
-  window['diff'](currentNode, newNode, body, document);
+  miso.diff(currentNode, newNode, body, document);
   expect(newNode.domRef.style['color']).toBe('');
 });
 
@@ -624,13 +626,13 @@ test('Should change css on a DOM node', () => {
   var currentNode = vnode('div', [], {}, {
     'color': 'red'
   });
-  window['diff'](null, currentNode, body, document)
+  miso.diff(null, currentNode, body, document)
 
   // test css change
   var newNode = vnode('div', [], {}, {
     'color': 'blue'
   });
-  window['diff'](currentNode, newNode, body, document);
+  miso.diff(currentNode, newNode, body, document);
   expect(newNode.domRef.style['color']).toBe('blue');
 });
 
@@ -642,13 +644,13 @@ test('Should no-op change to css on a DOM node', () => {
   var currentNode = vnode('div', [], {}, {
     'color': 'red'
   });
-  window['diff'](null, currentNode, body, document)
+  miso.diff(null, currentNode, body, document)
 
   // test css no-op change
   var newNode = vnode('div', [], {}, {
     'color': 'red'
   });
-  window['diff'](currentNode, newNode, body, document);
+  miso.diff(currentNode, newNode, body, document);
   expect(newNode.domRef.style['color']).toBe('red');
 });
 
@@ -664,10 +666,10 @@ test('Should call onCreated and onDestroyed', () => {
     destroy++;
   }, null, 'key');
 
-  window['diff'](null, currentNode, body, document)
+  miso.diff(null, currentNode, body, document)
   expect(create).toBe(1);
 
-  window['diff'](currentNode, null, body, document)
+  miso.diff(currentNode, null, body, document)
   expect(destroy).toBe(1);
 });
 
@@ -683,10 +685,10 @@ test('Should call onCreated and onBeforeDestroyed', () => {
     destroy++;
   }, 'key');
 
-  window['diff'](null, currentNode, body, document)
+  miso.diff(null, currentNode, body, document)
   expect(create).toBe(1);
 
-  window['diff'](currentNode, null, body, document)
+  miso.diff(currentNode, null, body, document)
   expect(destroy).toBe(1);
 });
 
@@ -701,8 +703,8 @@ test('Should call onDestroyed recursively', () => {
     }, null, 'a')], {}, {}, "html", null, null, function() {
       destroy++;
     }, null, 'b');
-  window['diff'](null, currentNode, body, document)
-  window['diff'](currentNode, null, body, document)
+  miso.diff(null, currentNode, body, document)
+  miso.diff(currentNode, null, body, document)
   expect(destroy).toBe(1);
   expect(childDestroy).toBe(1);
 });
@@ -719,13 +721,13 @@ test('Should call onBeforeDestroyed recursively', () => {
     }, 'a')], {}, {}, "html", null, null, null, function() {
       destroy++;
     }, 'b');
-  window['diff'](null, currentNode, body, document)
-  window['diff'](currentNode, null, body, document)
+  miso.diff(null, currentNode, body, document)
+  miso.diff(currentNode, null, body, document)
   expect(destroy).toBe(1);
   expect(childDestroy).toBe(1);
 });
 
-test('Should recreate a DOM node when tags are the same but keys are window different', () => {
+test('Should recreate a DOM node when tags are the same but keys are window miso.different', () => {
   var document = new jsdom.JSDOM().window.document;
   var body = document.body;
   var destroy = 0;
@@ -733,32 +735,32 @@ test('Should recreate a DOM node when tags are the same but keys are window diff
     vnode('div', [], {}, {}, "html", null, null, function() {
       destroy++;
     }, null, "key-1");
-  window['diff'](null, currentNode, body, document)
+  miso.diff(null, currentNode, body, document)
   var newNode =
     vnode('div', [], {}, {}, "html", null, null, function() {
       destroy++;
     }, null, "key-1");
-  window['diff'](null, currentNode, body, document)
+  miso.diff(null, currentNode, body, document)
   expect(destroy).toBe(0);
-  window['diff'](currentNode, newNode, body, document)
+  miso.diff(currentNode, newNode, body, document)
   var newKeyedNode =
     vnode('div', [], {}, {}, "html", null, null, function() {
       destroy++;
     }, null, "key-2");
-  window['diff'](currentNode, newKeyedNode, body, document)
+  miso.diff(currentNode, newKeyedNode, body, document)
   expect(destroy).toBe(1);
 });
 
-test('Should execute left-hand side happy path key-window diffing case', () => {
+test('Should execute left-hand side happy path key-window miso.diffing case', () => {
   var document = new jsdom.JSDOM().window.document;
   var body = document.body;
   var destroy = 0;
   var currentNode =
     vnode('div', [vnodeKeyed('div', 'a'), vnodeKeyed('div', 'b'), vnodeKeyed('div', 'c')], {}, {}, "html", null, null, null, null, "key-1");
-  window['diff'](null, currentNode, body, document)
+  miso.diff(null, currentNode, body, document)
   var newNode =
     vnode('div', [vnodeKeyed('div', 'a'), vnodeKeyed('div', 'b'), vnodeKeyed('div', 'c')], {}, {}, "html", null, null, null, null, "key-1");
-  window['diff'](currentNode, newNode, body, document)
+  miso.diff(currentNode, newNode, body, document)
   expect(newNode.children.length).toBe(3);
   expect(newNode.children.length).toBe(currentNode.children.length);
   expect(currentNode.children).toEqual(newNode.children);
@@ -766,16 +768,16 @@ test('Should execute left-hand side happy path key-window diffing case', () => {
   expect(currentNode.domRef.childNodes).toEqual(newNode.domRef.childNodes);
 });
 
-test('Should diff keys properly when keys are prepended', () => {
+test('Should miso.diff keys properly when keys are prepended', () => {
   var document = new jsdom.JSDOM().window.document;
   var body = document.body;
   var destroy = 0;
   var currentNode =
     vnode('div', [vnodeKeyed('div', '1')], {}, {}, "html", null, null, null, null, "key-1");
-  window['diff'](null, currentNode, body, document)
+  miso.diff(null, currentNode, body, document)
   var newNode =
     vnode('div', [vnodeKeyed('div', '2'), vnodeKeyed('div', '1')], {}, {}, "html", null, null, null, null, "key-1");
-  window['diff'](currentNode, newNode, body, document)
+  miso.diff(currentNode, newNode, body, document)
   expect(newNode.children.length).toBe(2);
   expect(newNode.children.length).toBe(currentNode.children.length);
   expect(currentNode.children).toEqual(newNode.children);
@@ -783,16 +785,16 @@ test('Should diff keys properly when keys are prepended', () => {
   expect(currentNode.domRef.childNodes).toEqual(newNode.domRef.childNodes);
 });
 
-test('Should execute right-hand side happy path key-window diffing case', () => {
+test('Should execute right-hand side happy path key-window miso.diffing case', () => {
   var document = new jsdom.JSDOM().window.document;
   var body = document.body;
   var destroy = 0;
   var currentNode =
     vnode('div', [vnodeKeyed('div', 'a'), vnodeKeyed('div', 'c')], {}, {}, "html", null, null, null, null, "key-1");
-  window['diff'](null, currentNode, body, document)
+  miso.diff(null, currentNode, body, document)
   var newNode =
     vnode('div', [vnodeKeyed('div', 'z'), vnodeKeyed('div', 'c')], {}, {}, "html", null, null, null, null, "key-1");
-  window['diff'](currentNode, newNode, body, document)
+  miso.diff(currentNode, newNode, body, document)
   expect(newNode.children.length).toBe(2);
   expect(newNode.children.length).toBe(currentNode.children.length);
   expect(currentNode.children).toEqual(newNode.children);
@@ -805,9 +807,9 @@ test('Should swap nodes', () => {
   var document = new jsdom.JSDOM().window.document;
   var body = document.body;
   var currentNode = vnode('div', [vnodeKeyed('div', 'a'), vnodeKeyed('div', 'b')], {}, {}, "html", null, null, null, null, "key-1");
-  window['diff'](null, currentNode, body, document)
+  miso.diff(null, currentNode, body, document)
   var newNode = vnode('div', [vnodeKeyed('div', 'b'), vnodeKeyed('div', 'a')], {}, {}, "html", null, null, null, null, "key-1");
-  window['diff'](currentNode, newNode, body, document)
+  miso.diff(currentNode, newNode, body, document)
   expect(newNode.children.length).toBe(2);
   expect(newNode.children.length).toBe(currentNode.children.length);
   expect(currentNode.children).toEqual(newNode.children);
@@ -820,10 +822,10 @@ test('Should execute flip-flop case', () => {
   var body = document.body;
   var currentNode =
     vnode('div', [vnodeKeyed('div', 'a'), vnodeKeyed('div', 'b'), vnodeKeyed('div', 'c')], {}, {}, "html", null, null, null, null, "key-1");
-  window['diff'](null, currentNode, body, document)
+  miso.diff(null, currentNode, body, document)
   var newNode =
     vnode('div', [vnodeKeyed('div', 'c'), vnodeKeyed('div', 'b'), vnodeKeyed('div', 'a')], {}, {}, "html", null, null, null, null, "key-1");
-  window['diff'](currentNode, newNode, body, document)
+  miso.diff(currentNode, newNode, body, document)
   expect(newNode.children.length).toBe(3);
   expect(newNode.children.length).toBe(currentNode.children.length);
   expect(currentNode.children).toEqual(newNode.children);
@@ -853,9 +855,9 @@ test('Should execute swapped case on 1k nodes', () => {
       newKids.push(vnodeKeyed('div', i))
     }
   }
-  window['diff'](null, currentNode, body, document)
+  miso.diff(null, currentNode, body, document)
   var newNode =  vnode('div', newKids, {}, {}, "html", null, null, null, null, "key-1");
-  window['diff'](currentNode, newNode, body, document)
+  miso.diff(currentNode, newNode, body, document)
   expect(newNode.children.length).toBe(1000);
   expect(newNode.children.length).toBe(currentNode.children.length);
   expect(currentNode.children).toEqual(newNode.children);
@@ -876,10 +878,10 @@ test('Should execute top-left and bottom-right match case', () => {
   var destroy = 0;
   var currentNode =
     vnode('div', [vnodeKeyed('div', 'd'), vnodeKeyed('div', 'a'), vnodeKeyed('div', 'k'), vnodeKeyed('div', 'r'), vnodeKeyed('div', 'b')], {}, {}, "html", null, null, null, null, "key-1");
-  window['diff'](null, currentNode, body, document)
+  miso.diff(null, currentNode, body, document)
   var newNode =
     vnode('div', [vnodeKeyed('div', 'a'), vnodeKeyed('div', 'b'), vnodeKeyed('div', 'r'), vnodeKeyed('div', 'k'), vnodeKeyed('div', 'd')], {}, {}, "html", null, null, null, null, "key-1");
-  window['diff'](currentNode, newNode, body, document)
+  miso.diff(currentNode, newNode, body, document)
   expect(newNode.children.length).toBe(5);
   expect(newNode.children.length).toBe(currentNode.children.length);
   expect(currentNode.children).toEqual(newNode.children);
@@ -893,10 +895,10 @@ test('Should handle duplicate keys case', () => {
   var destroy = 0;
   var currentNode =
     vnode('div', [vnodeKeyed('div', 'a'), vnodeKeyed('div', 'a'), vnodeKeyed('div', 'a'), vnodeKeyed('div', 'b'), vnodeKeyed('div', 'b')], {}, {}, "html", null, null, null, null, "key-1");
-  window['diff'](null, currentNode, body, document)
+  miso.diff(null, currentNode, body, document)
   var newNode =
     vnode('div', [vnodeKeyed('div', 'b'), vnodeKeyed('div', 'b'), vnodeKeyed('div', 'b'), vnodeKeyed('div', 'a'), vnodeKeyed('div', 'a')], {}, {}, "html", null, null, null, null, "key-1");
-  window['diff'](currentNode, newNode, body, document)
+  miso.diff(currentNode, newNode, body, document)
   expect(newNode.children.length).toBe(5);
   expect(newNode.children.length).toBe(currentNode.children.length);
   expect(currentNode.children).toEqual(newNode.children);
@@ -910,10 +912,10 @@ test('Should execute top-right and bottom-left match case', () => {
   var destroy = 0;
   var currentNode =
     vnode('div', [vnodeKeyed('div', 'd'), vnodeKeyed('div', 'a'), vnodeKeyed('div', 'g'), vnodeKeyed('div', 'b')], {}, {}, "html", null, null, null, null, "key-1");
-  window['diff'](null, currentNode, body, document)
+  miso.diff(null, currentNode, body, document)
   var newNode =
     vnode('div', [vnodeKeyed('div', 'b'), vnodeKeyed('div', 'g'), vnodeKeyed('div', 'd'), vnodeKeyed('div', 'a')], {}, {}, "html", null, null, null, null, "key-1");
-  window['diff'](currentNode, newNode, body, document)
+  miso.diff(currentNode, newNode, body, document)
   expect(newNode.children.length).toBe(4);
   expect(newNode.children.length).toBe(currentNode.children.length);
   expect(currentNode.children).toEqual(newNode.children);
@@ -927,10 +929,10 @@ test('Nothing matches case', () => {
   var destroy = 0;
   var currentNode =
     vnode('div', [vnodeKeyed('div', 'e'), vnodeKeyed('div', 'k'), vnodeKeyed('div', 'l')], {}, {}, "html", null, null, null, null, "key-1");
-  window['diff'](null, currentNode, body, document)
+  miso.diff(null, currentNode, body, document)
   var newNode =
     vnode('div', [vnodeKeyed('div', 'b'), vnodeKeyed('div', 'z'), vnodeKeyed('div', 'j')], {}, {}, "html", null, null, null, null, "key-1");
-  window['diff'](currentNode, newNode, body, document)
+  miso.diff(currentNode, newNode, body, document)
   expect(newNode.children.length).toBe(3);
   expect(newNode.children.length).toBe(currentNode.children.length);
   expect(currentNode.children).toEqual(newNode.children);
@@ -944,10 +946,10 @@ test('Should handle nothing matches case where new key is found in old map', () 
   var destroy = 0;
   var currentNode =
     vnode('div', [vnodeKeyed('div', 'a'), vnodeKeyed('div', 'k'), vnodeKeyed('div', 'l'), vnodeKeyed('div', 'c'), vnodeKeyed('div', 'g')], {}, {}, "html", null, null, null, "key-1");
-  window['diff'](null, currentNode, body, document)
+  miso.diff(null, currentNode, body, document)
   var newNode =
     vnode('div', [vnodeKeyed('div', 'b'), vnodeKeyed('div', 'c'), vnodeKeyed('div', 'l'), vnodeKeyed('div', 'r'), vnodeKeyed('div', 'k')], {}, {}, "html", null, null, null, "key-1");
-  window['diff'](currentNode, newNode, body, document)
+  miso.diff(currentNode, newNode, body, document)
   expect(newNode.children.length).toBe(5);
   expect(newNode.children.length).toBe(currentNode.children.length);
   expect(currentNode.children).toEqual(newNode.children);
@@ -961,10 +963,10 @@ test('Should append new nodes in keys patch', () => {
   var destroy = 0;
   var currentNode =
     vnode('div', [vnodeKeyed('div', 'a')], {}, {}, "html", null, null, null, "key-1");
-  window['diff'](null, currentNode, body, document)
+  miso.diff(null, currentNode, body, document)
   var newNode =
     vnode('div', [vnodeKeyed('div', 'a'), vnodeKeyed('div', 'c'), vnodeKeyed('div', 'k')], {}, {}, "html", null, null, null, "key-1");
-  window['diff'](currentNode, newNode, body, document)
+  miso.diff(currentNode, newNode, body, document)
   expect(newNode.children.length).toBe(3);
   expect(newNode.children.length).toBe(currentNode.children.length);
   expect(currentNode.children).toEqual(newNode.children);
@@ -972,14 +974,14 @@ test('Should append new nodes in keys patch', () => {
   expect(currentNode.domRef.childNodes).toEqual(newNode.domRef.childNodes);
 });
 
-test('Should window diff keyed text nodes', () => {
+test('Should window miso.diff keyed text nodes', () => {
   var document = new jsdom.JSDOM().window.document;
   var body = document.body;
   var destroy = 0;
   var currentNode = vnodeKids('div', [ vtextKeyed ("foo",1), vtextKeyed ("bar",2), vtextKeyed ("baz",3)]);
-  window['diff'](null, currentNode, body, document)
+  miso.diff(null, currentNode, body, document)
   var newNode = vnodeKids('div', [ vtextKeyed ("baz",3), vtextKeyed ("bar",2), vtextKeyed ("foo",1) ]);
-  window['diff'](currentNode, newNode, body, document);
+  miso.diff(currentNode, newNode, body, document);
   expect(newNode.children.length).toBe(currentNode.children.length);
   expect(newNode.children).toEqual(currentNode.children);
 });
@@ -994,7 +996,7 @@ test('Should copy simple nested DOM into VTree', () => {
   var txt = document.createTextNode("foo");
   nestedDiv.appendChild(txt);
   var currentNode = vnodeKids('div', [ vnodeKids('div', [ vtext("foo") ]) ]);
-  window['copyDOMIntoVTree'](true, body, currentNode, document);
+  miso.copyDOMIntoVTree(true, body, currentNode, document);
   expect(currentNode.children[0].children[0].text).toEqual('foo');
 });
 
@@ -1006,7 +1008,7 @@ test('Should fail because of expecting text node', () => {
   var nestedDiv = document.createElement("div");
   div.appendChild(nestedDiv);
   var currentNode = vnodeKids('div', [ vtext("foo") ]);
-  var res = window['copyDOMIntoVTree'](true, body, currentNode, document);
+  var res = miso.copyDOMIntoVTree(true, body, currentNode, document);
   expect(res).toEqual(false);
 });
 
@@ -1018,7 +1020,7 @@ test('Should fail because of expecting element', () => {
   var txt = document.createTextNode("foo");
   div.appendChild(txt);
   var currentNode = vnodeKids('div', [ vnode('div', []) ]);
-  var res = window['copyDOMIntoVTree'](true, body, currentNode, document);
+  var res = miso.copyDOMIntoVTree(true, body, currentNode, document);
   expect(res).toEqual(false);
 });
 
@@ -1030,7 +1032,7 @@ test('Should fail because of non-matching text', () => {
   var txt = document.createTextNode("foo");
   div.appendChild(txt);
   var currentNode = vnodeKids('div', [ vtext("bar") ]);
-  var res = window['copyDOMIntoVTree'](true, body, currentNode, document);
+  var res = miso.copyDOMIntoVTree(true, body, currentNode, document);
   expect(res).toEqual(false);
 });
 
@@ -1042,7 +1044,7 @@ test('Should fail because of non-matching DOM and VDOM', () => {
   var txt = document.createTextNode("foobar");
   div.appendChild(txt);
   var currentNode = vnodeKids('div', [ vtext("foo") ]);
-  var res = window['copyDOMIntoVTree'](true, body, currentNode, document);
+  var res = miso.copyDOMIntoVTree(true, body, currentNode, document);
   expect(res).toEqual(false);
 });
 
@@ -1054,7 +1056,7 @@ test('Should copy DOM into VTree with multiple consecutive text nodes and collap
   var txt = document.createTextNode("foobarbaz");
   div.appendChild(txt);
   var currentNode = vnodeKids('div', [ vtext("foo"), vtext("bar"), vtext("baz") ]);
-  window['copyDOMIntoVTree'](true, body, currentNode, document);
+  miso.copyDOMIntoVTree(true, body, currentNode, document);
   // Expect "foobarbaz" to be split up into three nodes in the DOM
   expect(div.childNodes[0].textContent).toEqual('foobarbaz');
 });
@@ -1067,7 +1069,7 @@ test('Should copy DOM into VTree with multiple consecutive text nodes and collap
   var txt = document.createTextNode("foobarbaz");
   div.appendChild(txt);
   var currentNode = vnodeKids('div', [ vtext("foo"), vtext("bar"), vtext("baz"), vnodeKids('div',[]), vtext("foo"), vtext("bar"), vtext("baz"), ]);
-  window['copyDOMIntoVTree'](true, null, currentNode, document);
+  miso.copyDOMIntoVTree(true, null, currentNode, document);
   // Expect "foobarbaz" to be split up into three nodes in the DOM
   expect(div.childNodes[0].textContent).toEqual('foobarbaz');
   expect(div.childNodes[2].textContent).toEqual('foobarbaz');
@@ -1091,7 +1093,7 @@ test('Should copy DOM into VTree at mountPoint', () => {
   var txt = document.createTextNode("foo");
   nestedDiv2.appendChild(txt);
   var currentNode = vnodeKids('div', [ vnodeKids('div', [ vtext("foo") ]) ]);
-  var succeeded = window['copyDOMIntoVTree'](true, misoDiv, currentNode, document);
+  var succeeded = miso.copyDOMIntoVTree(true, misoDiv, currentNode, document);
   expect(currentNode.children[0].children[0].domRef).toEqual(txt);
   expect(succeeded).toEqual(true);
 });
@@ -1114,7 +1116,7 @@ test('Should copy DOM into VTree at body w/ script / text siblings', () => {
   var txt = document.createTextNode("foo");
   nestedDiv2.appendChild(txt);
   var currentNode = vnodeKids('div', [ vnodeKids('div', [ vtext("foo") ]) ]);
-  var succeeded = window['copyDOMIntoVTree'](true, body, currentNode, document);
+  var succeeded = miso.copyDOMIntoVTree(true, body, currentNode, document);
   expect(currentNode.children[0].children[0].domRef).toEqual(txt);
   expect(succeeded).toEqual(false);
 });
@@ -1125,7 +1127,7 @@ test('Should fail to mount on a text node', () => {
   var misoTxt = document.createTextNode("foo");
   body.appendChild(misoTxt);
   var currentNode = vnodeKids('div', [ vnodeKids('div', [ vtext("foo") ]) ]);
-  var succeeded = window['copyDOMIntoVTree'](true, misoTxt, currentNode, document);
+  var succeeded = miso.copyDOMIntoVTree(true, misoTxt, currentNode, document);
   expect(succeeded).toEqual(false);
 });
 
@@ -1133,7 +1135,7 @@ test('Should mount on an empty body', () => {
   var document = new jsdom.JSDOM().window.document;
   var body = document.body;
   var currentNode = vnodeKids('div', [ vnodeKids('div', [ vtext("foo") ]) ]);
-  var succeeded = window['copyDOMIntoVTree'](true, null, currentNode, document);
+  var succeeded = miso.copyDOMIntoVTree(true, null, currentNode, document);
   expect(succeeded).toEqual(false);
 });
 
@@ -1154,9 +1156,9 @@ test('Should pass integrity check', () => {
     ns : 'HTML',
     css : {}
   };
-  var result = window['copyDOMIntoVTree'](false, body, vtree, document);
+  var result = miso.copyDOMIntoVTree(false, body, vtree, document);
   expect(result).toEqual(true);
-  var check = window['integrityCheck'](true, vtree);
+  var check = miso.integrityCheck(true, vtree);
   expect(check).toBe(1);
 });
 
@@ -1177,12 +1179,12 @@ test('Should fail integrity check on bad tag', () => {
     ns : 'HTML',
     css : {}
   };
-  var result = window['copyDOMIntoVTree'](false, body, vtree, document);
+  var result = miso.copyDOMIntoVTree(false, body, vtree, document);
   expect(result).toEqual(true);
-  var check = window['integrityCheck'](true, vtree);
+  var check = miso.integrityCheck(true, vtree);
   expect(check).toBe(1);
   vtree.tag = 'lol';
-  var check = window['integrityCheck'](true, vtree);
+  var check = miso.integrityCheck(true, vtree);
   expect(check).toBe(0);
 });
 
@@ -1203,11 +1205,11 @@ test('Should fail integrity check on bad tag in copyDOMIntoVTree w/ logging enab
      ns : 'HTML',
      css : {}
   };
-  var result = window['copyDOMIntoVTree'](true, body, vtree, document);
+  var result = miso.copyDOMIntoVTree(true, body, vtree, document);
   expect(result).toEqual(false);
 });
 
-test('Should fail integrity check on differing vtext', () => {
+test('Should fail integrity check on miso.differing vtext', () => {
   var document = new jsdom.JSDOM().window.document;
   var body = document.body;
   var child = document.createElement('div');
@@ -1224,16 +1226,16 @@ test('Should fail integrity check on differing vtext', () => {
     ns : 'HTML',
     css : {}
   };
-  var result = window['copyDOMIntoVTree'](false, body, vtree, document);
+  var result = miso.copyDOMIntoVTree(false, body, vtree, document);
   expect(result).toEqual(true);
-  var check = window['integrityCheck'](true, vtree);
+  var check = miso.integrityCheck(true, vtree);
   expect(check).toBe(1);
   vtree.children[0].text = 'oops';
-  var check = window['integrityCheck'](true, vtree);
+  var check = miso.integrityCheck(true, vtree);
   expect(check).toBe(0);
 });
 
-test('Should fail integrity check on differing child lengths', () => {
+test('Should fail integrity check on miso.differing child lengths', () => {
   var document = new jsdom.JSDOM().window.document;
   var body = document.body;
   var child = document.createElement('div');
@@ -1250,16 +1252,16 @@ test('Should fail integrity check on differing child lengths', () => {
      ns : 'HTML',
      css : {}
   };
-  var result = window['copyDOMIntoVTree'](false, body, vtree, document);
+  var result = miso.copyDOMIntoVTree(false, body, vtree, document);
   expect(result).toEqual(true);
-  var check = window['integrityCheck'](true, vtree);
+  var check = miso.integrityCheck(true, vtree);
   expect(check).toBe(1);
   vtree.children = [];
-  var check = window['integrityCheck'](true, vtree);
+  var check = miso.integrityCheck(true, vtree);
   expect(check).toBe(false);
 });
 
-test('Should fail integrity check on differing styles', () => {
+test('Should fail integrity check on miso.differing styles', () => {
   var document = new jsdom.JSDOM().window.document;
   var body = document.body;
   var child = document.createElement('div');
@@ -1277,16 +1279,16 @@ test('Should fail integrity check on differing styles', () => {
      ns : 'HTML',
      css : { 'background-color': 'red' }
   };
-  var result = window['copyDOMIntoVTree'](false, body, vtree, document);
+  var result = miso.copyDOMIntoVTree(false, body, vtree, document);
   expect(result).toEqual(true);
-  var check = window['integrityCheck'](true, vtree);
+  var check = miso.integrityCheck(true, vtree);
   expect(check).toBe(1);
   vtree.css['background-color'] = 'green';
-  var check = window['integrityCheck'](true, vtree);
+  var check = miso.integrityCheck(true, vtree);
   expect(check).toBe(0);
 });
 
-test('Should fail integrity check on differing styles, for color', () => {
+test('Should fail integrity check on miso.differing styles, for color', () => {
   var document = new jsdom.JSDOM().window.document;
   var body = document.body;
   var child = document.createElement('div');
@@ -1305,16 +1307,16 @@ test('Should fail integrity check on differing styles, for color', () => {
     ns : 'HTML',
     css : { 'background-color': 'red', color: '#cccccc' }
   };
-  var result = window['copyDOMIntoVTree'](false, body, vtree, document);
+  var result = miso.copyDOMIntoVTree(false, body, vtree, document);
   expect(result).toEqual(true);
-  var check = window['integrityCheck'](true, vtree);
+  var check = miso.integrityCheck(true, vtree);
   expect(check).toBe(1);
   vtree.css['color'] = '#dddddd';
-  var check = window['integrityCheck'](true, vtree);
+  var check = miso.integrityCheck(true, vtree);
   expect(check).toBe(0);
 });
 
-test('Should fail integrity check on differing props', () => {
+test('Should fail integrity check on miso.differing props', () => {
   var document = new jsdom.JSDOM().window.document;
   var body = document.body;
   var child = document.createElement('div');
@@ -1333,12 +1335,12 @@ test('Should fail integrity check on differing props', () => {
     ns : 'HTML',
     css : { 'background-color': 'red' }
   };
-  var result = window['copyDOMIntoVTree'](false, body, vtree, document);
+  var result = miso.copyDOMIntoVTree(false, body, vtree, document);
   expect(result).toEqual(true);
-  var check = window['integrityCheck'](true, vtree);
+  var check = miso.integrityCheck(true, vtree);
   expect(check).toBe(1);
   vtree.props['class'] = 'something-else';
-  var check = window['integrityCheck'](true, vtree);
+  var check = miso.integrityCheck(true, vtree);
   expect(check).toBe(0);
 });
 
@@ -1363,13 +1365,13 @@ test('Should fail integrity check on differing height / width', () => {
     ns : 'HTML',
     css : { 'background-color': 'red' }
   };
-  var result = window['copyDOMIntoVTree'](false, body, vtree, document);
+  var result = miso.copyDOMIntoVTree(false, body, vtree, document);
   expect(result).toEqual(true);
-  var check = window['integrityCheck'](true, vtree);
+  var check = miso.integrityCheck(true, vtree);
   expect(check).toBe(1);
   vtree.props['height'] = '200';
   vtree.props['width'] = '200';
-  var check = window['integrityCheck'](true, vtree);
+  var check = miso.integrityCheck(true, vtree);
   expect(check).toBe(0);
 });
 
@@ -1391,12 +1393,12 @@ test('Should fail integrity check on random property (title)', () => {
     ns : 'HTML',
     css : {}
   };
-  var result = window['copyDOMIntoVTree'](false, body, vtree, document);
+  var result = miso.copyDOMIntoVTree(false, body, vtree, document);
   expect(result).toEqual(true);
-  var check = window['integrityCheck'](true, vtree);
+  var check = miso.integrityCheck(true, vtree);
   expect(check).toBe(1);
   vtree.props['title'] = "woz";
-  var check = window['integrityCheck'](true, vtree);
+  var check = miso.integrityCheck(true, vtree);
   expect(check).toBe(0);
 });
 
@@ -1420,12 +1422,12 @@ test('Should fail integrity check on href', () => {
     ns : 'HTML',
     css : { 'background-color': 'red' }
   };
-  var result = window['copyDOMIntoVTree'](false, body, vtree, document);
+  var result = miso.copyDOMIntoVTree(false, body, vtree, document);
   expect(result).toEqual(true);
-  var check = window['integrityCheck'](true, vtree);
+  var check = miso.integrityCheck(true, vtree);
   expect(check).toBe(1);
   vtree.props['href'] = "notgoogle.com";
-  var check = window['integrityCheck'](true, vtree);
+  var check = miso.integrityCheck(true, vtree);
   expect(check).toBe(0);
 });
 
@@ -1448,12 +1450,12 @@ test('Should fail integrity check on vtext domRef', () => {
     ns : 'HTML',
     css : { 'background-color': 'red' }
   };
-  var result = window['copyDOMIntoVTree'](false, body, vtree, document);
+  var result = miso.copyDOMIntoVTree(false, body, vtree, document);
   expect(result).toEqual(true);
-  var check = window['integrityCheck'](true, vtree);
+  var check = miso.integrityCheck(true, vtree);
   expect(check).toBe(1);
   vtree.children[0].domRef = document.createElement('div');
-  var check = window['integrityCheck'](true, vtree);
+  var check = miso.integrityCheck(true, vtree);
   expect(check).toBe(0);
 });
 
@@ -1475,8 +1477,8 @@ test('Should fail integrity check on unknown property test', () => {
     ns : 'HTML',
     css : {}
   };
-  var result = window['copyDOMIntoVTree'](false, body, vtree, document);
+  var result = miso.copyDOMIntoVTree(false, body, vtree, document);
   expect(result).toEqual(true);
-  var check = window['integrityCheck'](true, vtree);
+  var check = miso.integrityCheck(true, vtree);
   expect(check).toBe(0);
 });
