+++ tests/package.json
@@ -2,10 +2,7 @@
   "name": "miso",
   "version": "1.0.0",
   "description": "miso tests",
-  "main": "diff.test.js",
-  "scripts": {
-    "test": "cp ../jsbits/diff.js ./ && cp ../jsbits/isomorphic.js ./ && echo 'module.exports = diff;' >> diff.js &&./node_modules/.bin/jest --collectCoverage=true && rm ./diff.js && rm ./isomorphic.js"
-  },
+  "main": "miso.spec.js",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/dmjio/miso.git"
